backend/requirements.txt
django==2.2
djangorestframework==3.9.2
#postgres
psycopg2>=2.7.5,<2.8.0
django-extensions==2.2.8
django-cors-headers==3.2.1
requests
BeautifulSoup4
django-rest-knox
gunicorn == 20.0.4